all component files styling extracted
