+++ src/app/components/admin-dashboard/navigation/admin-navigation-styles.tsx
@@ -2,7 +2,6 @@ import { makeStyles } from '@material-ui/core/styles';
 
 const drawerWidth = 240;
 export const useStyles = makeStyles((theme) => ({
-    
     drawer: {
         width: drawerWidth,
         flexShrink: 0,
