+++ src/app/components/child-dashboard/creative-content-submission/story-form/story-form.component.tsx
@@ -90,7 +90,6 @@ const StoryForm: React.FC<StoryFormProps> = ({ week, onUpdate }) => {
         ) {
             history.push('/kids-dashboard');
         }
-        console.log('eaeeeee');
     }, [submitted, removed, history, onUpdate, uploadedStory]);
 
     useEffect(() => {
