+++ src/app/components/dashboard/child/edit/edit-component-styles.tsx
@@ -1,4 +1,3 @@
-
 import { green } from '@material-ui/core/colors';
 import { makeStyles } from '@material-ui/core/styles';
