+++ src/app/components/dashboard/child/list/list.component.spec.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-
 import { ChildList } from './list.component';
 
 describe('ChildList', () => {
