+++ src/app/components/pdf-display/pdf-display/pdf-display-styles.tsx
@@ -1,0 +1,34 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'flex-start',
+        alignItems: 'center',
+    },
+    page: {
+        maxWidth: '100%',
+    },
+    orangeButton: {
+        'marginTop': '20px',
+        'backgroundColor': '#FF6B35',
+        'fontSize': '24px',
+        'fontWeight': 'bold',
+        'borderRadius': '50px',
+        'color': 'white',
+        'width': '200px',
+        'height': '50px',
+        'boxShadow': '0px 8px 0px #97300A',
+        'textTransform': 'capitalize',
+        'fontFamily': 'nunito',
+        '&:hover': {
+            backgroundColor: '#FF6B35',
+        },
+    },
+    btn: {
+        margin: '0 auto',
+        textAlign: 'center',
+        marginBottom: '30px',
+    },
+}));
