+++ src/app/components/point-allocation-dashboard/point-allocation-styles.ts
